# src/index.ts
import "dotenv/config";
import express from "express";
import swaggerUi from "swagger-ui-express";
import swaggerDocument from "../openapi.json" with { type: "json" };
import {
  findAll,
  findById,
  createTask,
  updateTask,
  deleteTask,
} from "./repositories/taskRepository.js";
import Task from "./types/task.js";
import supabase from "./supabase.js";
import { requireAuth } from "./middlewares/requireAuth.js";

const app = express();

app.use(express.json());
app.use("/docs", swaggerUi.serve, swaggerUi.setup(swaggerDocument));

app.get("/", (_req, res) => {
  res
    .status(200)
    .json({ name: "Task API", version: "1.0", endpoints: ["/tasks"] });
});

app.get("/tasks", async (_req, res) => {
  const tasks: Task[] = await findAll();

  res.status(200).json(tasks);
});

app.get("/tasks/:id", async (req, res) => {
  const id = Number(req.params.id);

  const task: Task | undefined = await findById(id);

  if (!task) {
    return res.status(404).json({
      error: `Task ${id} not found`,
    });
  }

  res.status(200).json(task);
});

app.post("/tasks", async (req, res) => {
  const { title } = req.body;

  if (typeof title !== "string" || title.trim() === "") {
    return res.status(400).json({
      error: "Title is required",
    });
  }

  const task: Task = await createTask(title);

  return res.status(201).json(task);
});

app.put("/tasks/:id", async (req, res) => {
  const id = Number(req.params.id);
  const { title, done } = req.body;

  if (title === undefined && done === undefined) {
    return res.status(400).json({
      error: "Nothing to update",
    });
  }

  if (title !== undefined) {
    if (typeof title !== "string" || title.trim() === "") {
      return res.status(400).json({
        error: "Title is required",
      });
    }
  }

  if (done !== undefined && typeof done !== "boolean") {
    return res.status(400).json({
      error: "Done must be a boolean",
    });
  }

  const task: Task | undefined = await updateTask(id, title, done);

  if (!task) {
    return res.status(404).json({
      error: `Task ${id} not found`,
    });
  }

  res.status(200).json(task);
});

app.delete("/tasks/:id", async (req, res) => {
  const id = Number(req.params.id);

  const deleted: boolean = await deleteTask(id);

  if (!deleted) {
    return res.status(404).json({
      error: `Task ${id} not found`,
    });
  }

  res.sendStatus(204);
});

app.post("/auth/signup", async (req, res) => {
  const { email, password } = req.body;

  if (
    typeof email !== "string" ||
    email.trim() === "" ||
    typeof password !== "string" ||
    password.trim() === ""
  ) {
    return res.status(400).json({
      error: "Email and password are required",
    });
  }

  const { data, error } = await supabase.auth.signUp({
    email,
    password,
  });

  if (error) {
    return res.status(400).json({
      error: error.message,
    });
  }

  return res.status(201).json({
    user: data.user,
  });
});

app.post("/auth/login", async (req, res) => {
  const { email, password } = req.body;

  if (
    typeof email !== "string" ||
    email.trim() === "" ||
    typeof password !== "string" ||
    password.trim() === ""
  ) {
    return res.status(400).json({
      error: "Email and password are required",
    });
  }

  const { data, error } = await supabase.auth.signInWithPassword({
    email,
    password,
  });

  if (error) {
    return res.status(401).json({
      error: "Invalid email or password",
    });
  }

  return res.status(200).json({
    accessToken: data.session?.access_token,
    refreshToken: data.session?.refresh_token,
    user: data.user,
  });
});

app.post("/auth/logout", requireAuth, async (_req, res) => {
  const { error } = await supabase.auth.signOut();

  if (error) {
    return res.status(400).json({
      error: error.message,
    });
  }

  return res.status(204).send();
});

app.get("/public/info", (_req, res) => {
  res.status(200).json({ message: "Welcome stranger! This info is public." });
});

app.get("/protected/profile", requireAuth, async (req, res) => {
  return res.json({
    id: req.user!.id,
    email: req.user!.email,
    createdAt: req.user!.created_at,
  });
});

app.get("/protected/dashboard", requireAuth, (req, res) => {
  res.json({
    message: `Welcome to your dashboard, ${req.user!.email}`,
  });
});

app.get("/health", (_req, res) => {
  res.status(200).json({
    status: "ok",
  });
});

const PORT = process.env.PORT || 3000;

app.listen(PORT, () => {
  console.log(`Server running on port ${PORT}`);
});